requirements.txt
# Core dependencies
streamlit>=1.24.0
tensorflow>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0

# TensorFlow Lite
tensorflow-lite-runtime>=2.13.0

# Web framework
streamlit>=1.24.0

# Utilities
pandas>=2.0.0
pathlib>=1.0.0

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# streamlit-aggrid>=0.3.0
# plotly>=5.15.0
# streamlit-lottie>=0.0.11
